+++ src/components/mediaPermission.js
@@ -3,7 +3,7 @@ import Swal from 'sweetalert2'
 import { phrases } from '../i18n'
 import { swalInfoOptions } from './swalOptions'
 
-import AllowCam from '../media/allow-camera.png?width=480&height=240'
+import AllowCam from '../media/allow-camera.png?width=240&height=120'
 
 export const checkPermissions = async RC => {
   if (navigator.permissions && navigator.permissions.query) {
@@ -17,7 +17,10 @@ export const checkPermissions = async RC => {
             imageUrl: AllowCam,
             imageWidth: 480,
             imageAlt: 'Please allow camera access',
-            html: phrases.RC_requestCamera[RC.L],
+            html:
+              phrases.RC_requestCamera[RC.L] +
+              '<br />' +
+              phrases.RC_privacyCamera[RC.L],
           })
         }
       })