+++ src/components/utils.js
@@ -126,14 +126,17 @@ export function isFullscreen() {
 export function constructInstructions(
   headline,
   description = null,
-  scrollable = false
+  scrollable = false,
+  descriptionClass = ''
 ) {
   return (
     `<div class="calibration-instruction${
       scrollable ? ' calibration-instruction-scrollable' : ''
     }"><h1>${headline}</h1>` +
     (description
-      ? `<p class="calibration-description">${description}</p></div>`
+      ? `<p class="calibration-description${
+          descriptionClass.length ? ' ' + descriptionClass : ''
+        }">${description}</p></div>`
       : '')
   )
 }
@@ -210,4 +213,13 @@ export const _copy = obj => {
  */
 export function blurAll() {
   if ('activeElement' in document) document.activeElement.blur()
+}
+
+/* -------------------------------------------------------------------------- */
+
+export const getClickOrTouchLocation = e => {
+  if (e.type === 'touchstart' || e.type === 'touchmove') {
+    const touch = e.touches[0]
+    return { x: touch.clientX, y: touch.clientY }
+  } else return { x: e.clientX, y: e.clientY }
 }