+++ src/gaze/gaze.js
@@ -19,7 +19,7 @@ RemoteCalibrator.prototype.trackGaze = async function (
    * showVideo: [Boolean]
    * pipWidthPx: [208]
    * showFaceOverlay: [Boolean]
-   * calibrationCount: [Number] Default 5
+   * calibrationCount: [Number] Default 1
    * decimalPlace: [Number] Default 0
    * thresholdDeg: [Number] or 'none'
    * headline: [String]
@@ -36,13 +36,13 @@ RemoteCalibrator.prototype.trackGaze = async function (
     {
       fullscreen: false,
       greedyLearner: false, // ! New 0.0.5
-      framerate: 30, // ! New 0.0.5
+      framerate: 60, // ! New 0.0.5
       showGazer: true,
       showVideo: true,
       pipWidthPx:
         this._CONST.N.VIDEO_W[this.isMobile.value ? 'MOBILE' : 'DESKTOP'],
       showFaceOverlay: false,
-      calibrationCount: 5,
+      calibrationCount: 1,
       thresholdDeg: 10, // minAccuracy
       decimalPlace: 0, // As the system itself has a high prediction error, it's not necessary to be too precise here
       headline: '👀 ' + phrases.RC_gazeTrackingTitle[this.L],
@@ -151,7 +151,10 @@ RemoteCalibrator.prototype.trackGaze = async function (
   }
 }
 
-RemoteCalibrator.prototype.getGazeNow = async function (callback = null) {
+RemoteCalibrator.prototype.getGazeNow = async function (
+  options = {},
+  callback = null
+) {
   if (
     !this.checkInitialized() ||
     !this.gazeTracker.checkInitialized('gaze', true) ||
@@ -160,9 +163,17 @@ RemoteCalibrator.prototype.getGazeNow = async function (callback = null) {
   )
     return
 
+  options = Object.assign(
+    {
+      wait: 0,
+      frames: 5,
+    },
+    options
+  )
+
   let c = callback || this.gazeTracker.defaultGazeCallback
 
-  return await this.gazeTracker.getGazeNow(c)
+  return await this.gazeTracker.getGazeNow(options, c)
 }
 
 RemoteCalibrator.prototype.pauseGaze = function () {